+++ src/tool/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 (This log only contains changes for grebedb-tool. Changes for the library are located in its own file.)
 
+## Unreleased
+
+* Added `inspect` command.
+
 ## 0.2.0 (2021-04-07)
 
 * Updated dependency to grebedb 0.2.0